Refactor: 메소드 별 파일 분리
- 메소드를 각각 GetAttemptsFromUser.js, GetCarNamesFromUser.js, GetResults.js 파일로 분리
