+++ src/constants.js
@@ -1,0 +1,6 @@
+export const GET_ATTEMPT_MESSAGE = "시도할 횟수는 몇 회인가요?";
+export const GET_CARS_MESSAGE = "경주할 자동차 이름을 입력하세요 (5자 이하): ";
+export const CARNAME_ERROR =
+	"[ERROR] 자동차 이름이 잘못된 형식입니다. (5자 이하만 가능)";
+export const ATTEMPT_NUMBER_ERROR = "[ERROR] 숫자 형식이 아닙니다.";
+export const WINNER_IS = "최종 우승자: ";
